+++ blockfrost-api/src/Blockfrost/Types/Cardano/Blocks.hs
@@ -27,6 +27,8 @@ data Block = Block
   , _blockOutput        :: Maybe Lovelaces -- ^ Total output within the block in Lovelaces
   , _blockFees          :: Maybe Lovelaces -- ^ Total fees within the block in Lovelaces
   , _blockBlockVrf      :: Maybe Text -- ^ VRF key of the block
+  , _blockOpCert        :: Maybe Text -- ^ The hash of the operational certificate of the block producer
+  , _blockOpCertCounter :: Maybe Quantity -- ^ The value of the counter used to produce the operational certificate
   , _blockPreviousBlock :: Maybe BlockHash -- ^ Hash of the previous block
   , _blockNextBlock     :: Maybe BlockHash -- ^ Hash of the next block
   , _blockConfirmations :: Integer -- ^ Number of block confirmations
@@ -49,6 +51,8 @@ instance ToSample Block where
     , _blockOutput = pure 128314491794
     , _blockFees = pure 592661
     , _blockBlockVrf = pure "vrf_vk1wf2k6lhujezqcfe00l6zetxpnmh9n6mwhpmhm0dvfh3fxgmdnrfqkms8ty"
+    , _blockOpCert = pure "da905277534faf75dae41732650568af545134ee08a3c0392dbefc8096ae177c"
+    , _blockOpCertCounter = pure 18
     , _blockPreviousBlock = pure "43ebccb3ac72c7cebd0d9b755a4b08412c9f5dcb81b8a0ad1e3c197d29d47b05"
     , _blockNextBlock = pure "8367f026cf4b03e116ff8ee5daf149b55ba5a6ec6dec04803b8dc317721d15fa"
     , _blockConfirmations = 4698