+++ blockfrost-api/src/Blockfrost/Types/Cardano/Epochs.hs
@@ -4,14 +4,23 @@ module Blockfrost.Types.Cardano.Epochs
   ( EpochInfo (..)
   , PoolStakeDistribution (..)
   , ProtocolParams (..)
+  , CostModels (..)
   , StakeDistribution (..)
   ) where
 
 import Blockfrost.Types.Shared
-import Data.Aeson (Value)
+import Data.Aeson (object, FromJSON (..), ToJSON (..), withObject)
+import Data.Map (Map)
 import Data.Text (Text)
 import Deriving.Aeson
 import Servant.Docs (ToSample (..), singleSample)
+
+import qualified Data.Aeson.Key
+import qualified Data.Aeson.KeyMap
+import qualified Data.Char
+import qualified Data.Map
+
+import Blockfrost.Types.Cardano.Scripts (ScriptType (..))
 
 -- | Information about an epoch
 data EpochInfo = EpochInfo
@@ -45,7 +54,6 @@ instance ToSample EpochInfo where
       , _epochInfoActiveStake = pure 784953934049314
       }
 
-
 -- | Protocol parameters
 data ProtocolParams = ProtocolParams
   { _protocolParamsEpoch                 :: Epoch -- ^ Epoch number
@@ -58,20 +66,19 @@ data ProtocolParams = ProtocolParams
   , _protocolParamsPoolDeposit           :: Lovelaces -- ^ The amount of a pool registration deposit in Lovelaces
   , _protocolParamsEMax                  :: Integer -- ^ Epoch bound on pool retirement
   , _protocolParamsNOpt                  :: Integer -- ^ Desired number of pools
-  , _protocolParamsA0                    :: Double -- ^ Pool pledge influence
-  , _protocolParamsRho                   :: Double -- ^ Monetary expansion
-  , _protocolParamsTau                   :: Double -- ^ Treasury expansion
-  , _protocolParamsDecentralisationParam :: Double -- ^ Percentage of blocks produced by federated nodes
-  , _protocolParamsExtraEntropy          :: Maybe Value -- ^ Seed for extra entropy
+  , _protocolParamsA0                    :: Rational -- ^ Pool pledge influence
+  , _protocolParamsRho                   :: Rational -- ^ Monetary expansion
+  , _protocolParamsTau                   :: Rational -- ^ Treasury expansion
+  , _protocolParamsDecentralisationParam :: Rational -- ^ Percentage of blocks produced by federated nodes
+  , _protocolParamsExtraEntropy          :: Maybe Text -- ^ Seed for extra entropy
   , _protocolParamsProtocolMajorVer      :: Integer -- ^ Accepted protocol major version
   , _protocolParamsProtocolMinorVer      :: Integer -- ^ Accepted protocol minor version
   , _protocolParamsMinUtxo               :: Lovelaces -- ^ Minimum UTXO value
   , _protocolParamsMinPoolCost           :: Lovelaces  -- ^ Minimum stake cost forced on the pool
   , _protocolParamsNonce                 :: Text -- ^ Epoch number only used once
-  -- cost models
-  -- https://github.com/input-output-hk/cardano-db-sync/pull/758
-  , _protocolParamsPriceMem               :: Double -- ^ The per word cost of script memory usage
-  , _protocolParamsPriceStep              :: Double -- ^ The cost of script execution step usage
+  , _protocolParamsCostModels            :: CostModels -- ^ Cost models parameters for Plutus Core scripts
+  , _protocolParamsPriceMem               :: Rational -- ^ The per word cost of script memory usage
+  , _protocolParamsPriceStep              :: Rational -- ^ The cost of script execution step usage
   , _protocolParamsMaxTxExMem             :: Quantity -- ^ The maximum number of execution memory allowed to be used in a single transaction
   , _protocolParamsMaxTxExSteps           :: Quantity -- ^ The maximum number of execution steps allowed to be used in a single transaction
   , _protocolParamsMaxBlockExMem          :: Quantity -- ^ The maximum number of execution memory allowed to be used in a single block
@@ -79,7 +86,8 @@ data ProtocolParams = ProtocolParams
   , _protocolParamsMaxValSize             :: Quantity -- ^ The maximum Val size
   , _protocolParamsCollateralPercent      :: Integer -- ^ The percentage of the transactions fee which must be provided as collateral when including non-native scripts
   , _protocolParamsMaxCollateralInputs    :: Integer -- ^ The maximum number of collateral inputs allowed in a transaction
-  , _protocolParamsCoinsPerUtxoWord       :: Lovelaces -- ^ The cost per UTxO word
+  , _protocolParamsCoinsPerUtxoSize       :: Lovelaces -- ^ The cost per UTxO size. Cost per UTxO *word* for Alozno. Cost per UTxO *byte* for Babbage and later
+  , _protocolParamsCoinsPerUtxoWord       :: Lovelaces -- ^ The cost per UTxO word (DEPRECATED)
   }
   deriving stock (Show, Eq, Generic)
   deriving (FromJSON, ToJSON)
@@ -108,6 +116,7 @@ instance ToSample ProtocolParams where
       , _protocolParamsMinUtxo = 1000000
       , _protocolParamsMinPoolCost = 340000000
       , _protocolParamsNonce = "1a3be38bcbb7911969283716ad7aa550250226b76a61fc51cc9a9a35d9276d81"
+      , _protocolParamsCostModels = costModelsSample
       , _protocolParamsPriceMem = 0.0577
       , _protocolParamsPriceStep = 0.0000721
       , _protocolParamsMaxTxExMem = 10000000
@@ -117,8 +126,65 @@ instance ToSample ProtocolParams where
       , _protocolParamsMaxValSize = 5000
       , _protocolParamsCollateralPercent = 150
       , _protocolParamsMaxCollateralInputs = 3
+      , _protocolParamsCoinsPerUtxoSize = 34482
       , _protocolParamsCoinsPerUtxoWord = 34482
       }
+
+newtype CostModels = CostModels { unCostModels :: Map ScriptType (Map Text Integer) }
+  deriving (Eq, Show, Generic)
+
+instance ToJSON CostModels where
+  toJSON =
+      object
+    . map (\(lang, params) ->
+        ( Data.Aeson.Key.fromString $ show lang
+        , object
+          $ map (\(key, param) ->
+              ( Data.Aeson.Key.fromText key
+              , toJSON param)
+            )
+            $ Data.Map.toList params
+        ))
+    . Data.Map.toList
+    . unCostModels
+
+instance FromJSON CostModels where
+  parseJSON = withObject "CostModel" $ \o -> do
+    let parseParams = withObject "CostModelParams" $ \po -> do
+          mapM (parseJSON . toJSON) $ Data.Aeson.KeyMap.toList po
+
+    langs <- mapM
+               (\(kLang, vParams) -> do
+                 l <- parseJSON
+                    $ toJSON
+                    $ (\(x:xs) -> Data.Char.toLower x:xs)
+                    $ Data.Aeson.Key.toString kLang
+                 ps <- parseParams vParams
+                 pure (l, Data.Map.fromList ps)
+               )
+               $ Data.Aeson.KeyMap.toList o
+
+    pure $ CostModels $ Data.Map.fromList langs
+
+costModelsSample :: CostModels
+costModelsSample = CostModels
+      $ Data.Map.fromList
+      [ ( PlutusV1
+        , Data.Map.fromList
+          [ ("addInteger-cpu-arguments-intercept", 197209)
+          , ("addInteger-cpu-arguments-slope", 0)
+          ]
+        )
+      , (PlutusV2
+        , Data.Map.fromList
+          [ ("addInteger-cpu-arguments-intercept", 197209)
+          , ("addInteger-cpu-arguments-slope", 0)
+          ]
+        )
+      ]
+
+instance ToSample CostModels where
+  toSamples = pure $ singleSample costModelsSample
 
 -- | Active stake distribution for an epoch
 data StakeDistribution = StakeDistribution