+++ blockfrost-api/test/Cardano/Pools.hs
@@ -86,6 +86,7 @@ poolInfoSample = [r|
   "hex": "0f292fcaa02b8b2f9b3c8f9fd8e0bb21abedb692a6d5058df3ef2735",
   "vrf_key": "0b5245f9934ec2151116fb8ec00f35fd00e0aa3b075c4ed12cce440f999d8233",
   "blocks_minted": 69,
+  "blocks_epoch": 4,
   "live_stake": "6900000000",
   "live_size": 0.42,
   "live_saturation": 0.93,
@@ -117,6 +118,7 @@ poolInfoExpected =
     , _poolInfoHex = "0f292fcaa02b8b2f9b3c8f9fd8e0bb21abedb692a6d5058df3ef2735"
     , _poolInfoVrfKey = "0b5245f9934ec2151116fb8ec00f35fd00e0aa3b075c4ed12cce440f999d8233"
     , _poolInfoBlocksMinted = 69
+    , _poolInfoBlocksEpoch = 4
     , _poolInfoLiveStake = 6900000000
     , _poolInfoLiveSize = 0.42
     , _poolInfoLiveSaturation = 0.93