+++ blockfrost-client/src/Blockfrost/Client/Cardano/Addresses.hs
@@ -36,11 +36,15 @@ getAddressUtxos_ :: MonadBlockfrost m => Project -> Address -> Paged -> SortOrde
 getAddressUtxos_ = _addressUtxos . addressesClient
 
 -- | UTXOs of the address.
--- Allows custom paging and ordering using @Paged@ and @SortOrder@.
+-- Allows custom paging and ordering using 'Paged' and 'SortOrder'.
 getAddressUtxos' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AddressUtxo]
 getAddressUtxos' a pg s = go (\p -> getAddressUtxos_ p a pg s)
 
 -- | UTXOs of the address.
+--
+-- Queries 100 entries. To query all entries use 'Blockfrost.Client.Core.allPages'
+-- with principled variant of this function (suffixed with @'@)
+-- that accepts 'Paged' argument.
 getAddressUtxos :: MonadBlockfrost m => Address -> m [AddressUtxo]
 getAddressUtxos a = getAddressUtxos' a def def
 
@@ -48,11 +52,15 @@ getAddressUtxosAsset_ :: MonadBlockfrost m => Project -> Address -> AssetId -> P
 getAddressUtxosAsset_ = _addressUtxosAsset . addressesClient
 
 -- | UTXOs of the address containing specific asset.
--- Allows custom paging and ordering using @Paged@ and @SortOrder@.
+-- Allows custom paging and ordering using 'Paged' and 'SortOrder'.
 getAddressUtxosAsset' :: MonadBlockfrost m => Address -> AssetId-> Paged -> SortOrder -> m [AddressUtxo]
 getAddressUtxosAsset' addr asset pg s = go (\p -> getAddressUtxosAsset_ p addr asset pg s)
 
 -- | UTXOs of the address containing specific asset.
+--
+-- Queries 100 entries. To query all entries use 'Blockfrost.Client.Core.allPages'
+-- with principled variant of this function (suffixed with @'@)
+-- that accepts 'Paged' argument.
 getAddressUtxosAsset :: MonadBlockfrost m => Address -> AssetId -> m [AddressUtxo]
 getAddressUtxosAsset addr asset = getAddressUtxosAsset' addr asset def def
 
@@ -68,7 +76,7 @@ getAddressTransactions_ ::
 getAddressTransactions_ = _addressTransactions . addressesClient
 
 -- | Transactions on the address.
--- Allows custom paging and ordering using @Paged@ and @SortOrder@.
+-- Allows custom paging and ordering using 'Paged' and 'SortOrder'.
 -- Also allows support for limiting block ranges using `from`/`to`
 -- @BlockIndex@es.
 getAddressTransactions' ::
@@ -82,5 +90,9 @@ getAddressTransactions' ::
 getAddressTransactions' a pg s from to = go (\p -> getAddressTransactions_ p a pg s from to)
 
 -- | Transactions on the address.
+--
+-- Queries 100 entries. To query all entries use 'Blockfrost.Client.Core.allPages'
+-- with principled variant of this function (suffixed with @'@)
+-- that accepts 'Paged' argument.
 getAddressTransactions :: MonadBlockfrost m => Address -> m [AddressTransaction]
 getAddressTransactions a = getAddressTransactions' a def def Nothing Nothing